% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCLLPS.R
\name{predict_GCLLPS}
\alias{predict_GCLLPS}
\title{Calculate Gastric Cancer Lifespan LncRNA Prognostic Signature (GCLLPS)}
\usage{
predict_GCLLPS(my_expr)
}
\arguments{
\item{my_expr}{A numeric matrix with rows as samples and columns as genes.}
}
\value{
A data frame with sample names and their GCLLPS scores.
}
\description{
This function calculates the GCLLPS based on a given gene expression matrix.
The expression matrix should have rows as samples and columns as genes.
If the matrix is missing required genes, an error will be thrown.
}
